♻️  Refactor create user
